Fixed 1D plotting for edge cases
It was possible to set `max_cols` to be less than the number of parameters to
plot, in which case the output would just be a single axis.
Furthermore, with `max_cols` in place, the output of `gs.subplots` wasn't a
flat numpy array, but rather a matrix, which messed up the getter for
`FisherFigure1D`.
Finally, if we plotted only one parameter, `gs.subplots` would just return the
`Axes` object, and not a numpy array with that object, which messed up
`plot_1d` itself.