Fixed placement of legend in 2D plot
This was annoyingly difficult to solve: apparently, matplotlib does not
support removing the legend from the figure, so multiple calls to
`figure.legend` would just draw on top of the existing one. On the other
hand, while the axes support removing the legend, calling `axes.legend`
with the same `bbox_to_anchor` coords does not produce the desired
results, because the coords are now relative to the axis, not the
figure! Fortunately, I stumbled upon this SO answer:
https://stackoverflow.com/a/61026525
which explains the `bbox_transform` argument, and allows one to place
the legend in the same location as before.