Added `validate_parameter` method to `FisherDerivative`
Fixes #17.
It's possible that a user makes a typo on some parameter when calling
`fisher_matrix`, and therefore the whole computation may fail, wasting
valuable time. This commit adds a method called `validate_parameter`,
which is meant to be overridden if the user wants `fisher_matrix` to
raise a `ValidationError` if an invalid parameter is encountered.