Updated test images and tests
It appears that Matplotlib 3.6 changed something when computing the
bounds of a figure when calling `savefig` with the `bbox_inches='tight'`
argument, causing the files to have differing dimensions (by a pixel or
so) between MPL 3.5 and 3.6. As a result, I've regenerated the figures
removing that argument, and setting the diff threshold to 20, as there
does not appear to be a discernable difference (to the human eye, at
least).