Replaced `covariance` arg with `inverse_covariance`
The method `matrix_element_from_input` used to take a `covariance`
argument. Unfortunately, when calling the method repeatedly, the
constant inverting of the matrix becomes a bit of a bottleneck, so it is
much more performant to do the inversion once, and then pass this
inverse to the function instead.