Added option to use pseudoinverse
Added `use_pinv` kwarg (defaults to False) to various functions that use
the matrix inverse, which toggles whether to use the Moore-Penrose
pesudoinverse instead of attempting to compute the regular inverse. This
should be used with caution, as, while the Moore-Penrose pseudoinverse
is unique, the results may not be completely reliable.